python/requirements.txt
pyserial==3.4
pywin32==227
